Separate DescriptionList filter enums into their own files.
- Refactoring to make DescriptionList a smaller file.
